refactor: 팀 보따리 현황 ui모델 클래스 이동
