+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/model/MemberCheckStatusUiModel.kt
@@ -1,0 +1,6 @@
+package com.bottari.presentation.model
+
+data class MemberCheckStatusUiModel(
+    val name: String,
+    val checked: Boolean,
+)
