+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/model/TeamBottariStatusUiModel.kt
@@ -1,0 +1,6 @@
+package com.bottari.presentation.model
+
+data class TeamBottariStatusUiModel(
+    val sharedItems: List<TeamProductStatusUiModel>,
+    val assignedItems: List<TeamProductStatusUiModel>,
+)
